Sound change

+++ src/main/java/dev/kyro/pitsim/controllers/OutpostManager.java
@@ -49,7 +49,7 @@ public class OutpostManager implements Listener {
 				strings.add("&3&lOUTPOST");
 				strings.add("&8&m------------------");
 				strings.add("&7Held by: &b" + (controllingGuild == null ? "&cNone" : controllingGuild.color + controllingGuild.name));
-				strings.add("&7Control: &f" + percentControlled + "%");
+				strings.add("&7Control: &f" + (int) percentControlled + "%");
 				strings.add("&7Status: " + getStatus());
 				strings.add("&8&m------------------");
 				strings.add("&7Rewards &6+" + Formatter.formatGoldFull(getGoldAmount()) + "&7/10m, &d+" + getOutpostFreshIncrease() + "% Mystic Find");
